WBackendVlc: Update OpenGL functions to integer #3

+++ src/SkMedia/src/media/WBackendVlc_p.h
@@ -154,10 +154,10 @@ public: // Variables
 
     QRect targetRect;
 
-    int targetX;
-    int targetY;
-    int targetWidth;
-    int targetHeight;
+    GLfloat targetX;
+    GLfloat targetY;
+    GLfloat targetWidth;
+    GLfloat targetHeight;
 
     bool shader;
